requirements.txt
anyio==3.6.1
autopep8==1.6.0
cachetools==5.2.0
certifi==2022.6.15
chardet==3.0.4
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
fastapi==0.79.0
h11==0.13.0
idna==2.6
imageio==2.19.3
networkx==2.8.4
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydantic==1.9.1
pyhumps==3.7.2
pymongo==4.1.1
pyparsing==3.0.9
python-multipart==0.0.5
PyWavelets==1.3.0
rsa==4.8
scikit-image==0.19.3
scipy==1.8.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tifffile==2022.5.4
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.22
uvicorn==0.18.2